+++ mycodo/mycodo_flask/app.py
@@ -2,9 +2,9 @@
 #
 #  app.py - Flask web server for Mycodo
 #
+import base64
 import logging
 
-import base64
 import flask_login
 from flask import Flask
 from flask import flash
@@ -40,6 +40,7 @@ from mycodo.mycodo_flask import routes_static
 from mycodo.mycodo_flask.api import api_blueprint
 from mycodo.mycodo_flask.api import init_api
 from mycodo.mycodo_flask.extensions import db
+from mycodo.mycodo_flask.utils.utils_dashboard import add_widget_endpoints
 from mycodo.mycodo_flask.utils.utils_general import get_ip_address
 
 logger = logging.getLogger(__name__)
@@ -58,6 +59,8 @@ def create_app(config=ProdConfig):
 
     register_extensions(app)
     register_blueprints(app)
+
+    add_widget_endpoints(app)
 
     return app